+++ storage/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+* feat: `Delete` may return `ObjectNotFound` [#425](https://github.com/Scalingo/go-utils/pull/425)
+
 ## v1.3.1
 
 * feat(list): `List` can take a `ListOpts` which can be used to define a `MaxKeys` field. This field limits the amount of objects in the returned list. By default this variable equals 0 which means that S3 will return the maximum objects he can (max is storage.S3ListMaxKeys = 1,000) [#427](https://github.com/Scalingo/go-utils/pull/427)